client, kv: split the code in multiple files
... to ease reading and maintenance.

Release note: None
