+++ pkg/kv/txn_coord_sender_factory.go
@@ -1,0 +1,102 @@
+// Copyright 2019 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package kv
+
+import (
+	"time"
+
+	"github.com/cockroachdb/cockroach/pkg/base"
+	"github.com/cockroachdb/cockroach/pkg/internal/client"
+	"github.com/cockroachdb/cockroach/pkg/roachpb"
+	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
+	"github.com/cockroachdb/cockroach/pkg/util/hlc"
+	"github.com/cockroachdb/cockroach/pkg/util/log"
+	"github.com/cockroachdb/cockroach/pkg/util/metric"
+	"github.com/cockroachdb/cockroach/pkg/util/stop"
+)
+
+// TxnCoordSenderFactory implements client.TxnSenderFactory.
+type TxnCoordSenderFactory struct {
+	log.AmbientContext
+
+	st                *cluster.Settings
+	wrapped           client.Sender
+	clock             *hlc.Clock
+	heartbeatInterval time.Duration
+	linearizable      bool // enables linearizable behavior
+	stopper           *stop.Stopper
+	metrics           TxnMetrics
+
+	testingKnobs ClientTestingKnobs
+}
+
+var _ client.TxnSenderFactory = &TxnCoordSenderFactory{}
+
+// TxnCoordSenderFactoryConfig holds configuration and auxiliary objects that can be passed
+// to NewTxnCoordSenderFactory.
+type TxnCoordSenderFactoryConfig struct {
+	AmbientCtx log.AmbientContext
+
+	Settings *cluster.Settings
+	Clock    *hlc.Clock
+	Stopper  *stop.Stopper
+
+	HeartbeatInterval time.Duration
+	Linearizable      bool
+	Metrics           TxnMetrics
+
+	TestingKnobs ClientTestingKnobs
+}
+
+// NewTxnCoordSenderFactory creates a new TxnCoordSenderFactory. The
+// factory creates new instances of TxnCoordSenders.
+func NewTxnCoordSenderFactory(
+	cfg TxnCoordSenderFactoryConfig, wrapped client.Sender,
+) *TxnCoordSenderFactory {
+	tcf := &TxnCoordSenderFactory{
+		AmbientContext:    cfg.AmbientCtx,
+		st:                cfg.Settings,
+		wrapped:           wrapped,
+		clock:             cfg.Clock,
+		stopper:           cfg.Stopper,
+		linearizable:      cfg.Linearizable,
+		heartbeatInterval: cfg.HeartbeatInterval,
+		metrics:           cfg.Metrics,
+		testingKnobs:      cfg.TestingKnobs,
+	}
+	if tcf.st == nil {
+		tcf.st = cluster.MakeTestingClusterSettings()
+	}
+	if tcf.heartbeatInterval == 0 {
+		tcf.heartbeatInterval = base.DefaultTxnHeartbeatInterval
+	}
+	if tcf.metrics == (TxnMetrics{}) {
+		tcf.metrics = MakeTxnMetrics(metric.TestSampleInterval)
+	}
+	return tcf
+}
+
+// TransactionalSender is part of the TxnSenderFactory interface.
+func (tcf *TxnCoordSenderFactory) TransactionalSender(
+	typ client.TxnType, meta roachpb.TxnCoordMeta, pri roachpb.UserPriority,
+) client.TxnSender {
+	return newTxnCoordSender(tcf, typ, meta, pri)
+}
+
+// NonTransactionalSender is part of the TxnSenderFactory interface.
+func (tcf *TxnCoordSenderFactory) NonTransactionalSender() client.Sender {
+	return tcf.wrapped
+}
+
+// Metrics returns the factory's metrics struct.
+func (tcf *TxnCoordSenderFactory) Metrics() TxnMetrics {
+	return tcf.metrics
+}
